move audioPlayer to hooks, rename to useAudioPlayer, presave useAudioPlayer for functionality refactor
