+++ components/Player.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
+import { useAudioPlayer } from '../hooks/useAudioPlayer';
 
 // test  creation of audio through react web player
 
 const Player = () => {
+  const audio = useAudioPlayer();
   return <div></div>;
 };
